// 02-nest-restful-api/src/model/user.model.ts
export class RegisterUserRequest {
  username: string;
  password: string;
  name: string;
}

export class LoginUserRequest {
  username: string;
  password: string;
}

export class UpdateUserRequest {
  password?: string;
  name?: string;
}

export class UserResponse {
  username: string;
  name: string;
  token?: string;
}